+++ runtime/include/tasks/fifo/chpl-tasks-impl-fns.h
@@ -1,0 +1,89 @@
+/*
+ * Copyright 2020 Hewlett Packard Enterprise Development LP
+ * Copyright 2004-2019 Cray Inc.
+ * Other additional copyright holders may be indicated within.
+ * 
+ * The entirety of this work is licensed under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * 
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+//
+// Fifo tasking layer interface function shortcuts.
+//
+
+#ifndef _chpl_tasks_impl_fns_h_
+#define _chpl_tasks_impl_fns_h_
+
+#include <stdint.h>
+
+#include "chpl-tasks-prvdata.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+//
+// Because we use the task tracking table for fifo tasking, this gives
+// us the ability to register the main task in the task table and
+// block report.
+//
+#define CHPL_TASK_STD_MODULES_INITIALIZED chpl_task_stdModulesInitialized
+void chpl_task_stdModulesInitialized(void);
+
+
+//
+// The fifo tasking layer doesn't really support sublocales.
+//
+// Putting these interface function definitions here and marking them
+// for inlining makes them cost-free at execution time.
+//
+#ifdef CHPL_TASK_GETSUBLOC_IMPL_DECL
+#error "CHPL_TASK_GETSUBLOC_IMPL_DECL is already defined!"
+#else
+#define CHPL_TASK_GETSUBLOC_IMPL_DECL 1
+#endif
+static inline
+c_sublocid_t chpl_task_getSubloc(void) {
+  return 0;
+}
+
+
+#ifdef CHPL_TASK_SETSUBLOC_IMPL_DECL
+#error "CHPL_TASK_SETSUBLOC_IMPL_DECL is already defined!"
+#else
+#define CHPL_TASK_SETSUBLOC_IMPL_DECL 1
+#endif
+static inline
+void chpl_task_setSubloc(c_sublocid_t subloc) {
+  // nothing to do
+}
+
+
+#ifdef CHPL_TASK_GETREQUESTEDSUBLOC_IMPL_DECL
+#error "CHPL_TASK_GETREQUESTEDSUBLOC_IMPL_DECL is already defined!"
+#else
+#define CHPL_TASK_GETREQUESTEDSUBLOC_IMPL_DECL 1
+#endif
+static inline
+c_sublocid_t chpl_task_getRequestedSubloc(void) {
+  return c_sublocid_any;
+}
+
+#define CHPL_TASK_IMPL_CAN_MIGRATE_THREADS() 0
+
+#ifdef __cplusplus
+} // end extern "C"
+#endif
+
+#endif
